chore(backend/navigate): apply navigation logic to singleplayer navigation aswell
